osscs/requirements.txt
Django==4.0.2
djangorestframework==3.13.1
drf_spectacular==0.21.2
requests==2.27.1
uwsgi==2.0.20
drf-spectacular-sidecar==2022.2.7